Added AFK placeholder + Translations in en-ES.yml.
This adds a placeholder called "%afk" that can be used for example to show the players that are AFK in the tablist. I have put an option to disable this placeholder (and the integration with Essentials, which is needed for this). By default it is disabled.

To activate it, just set in the config "enable" to "true" in AfkPlaceholder. Once this is done we can use the variable in our tablist by adding it in format (Example: %afk%prefix%displayname%suffix).

I have also translated some things in the en-ES.yml file.